Cached the S3 client
Previously if the host app was using the presigning functionality, we
were continously opening the S3 client which caused an N1 issue. For
example if the host was presigning 100 files, thats 100 new S3 clients
which queries the internal Ec2 Metadata service 100 times for creds. The
issue is that the Ec2 metadata service starts the throttle you if you
hit it too many times.

In this commit we create a new S3 object on boot and cache it in memory